improve for syntax
commenting
